# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRD Engineering Gap Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Gap {
  id: number;
  analysisId: number;
  gapType: string;
  description: string;
  severity: string;
  confidence: number;
  createdAt: string;
  helpfulCount: number;
  notHelpfulCount: number;
}

export interface Analysis {
  id: number;
  prdText: string;
  title: string;
  createdAt: string;
  gaps: Gap[];
}

export interface AnalysisSummary {
  id: number;
  title: string;
  prdText: string;
  createdAt: string;
  gapCount: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
}

export interface AnalysisInput {
  /** @minLength 50 */
  prdText: string;
}

export interface GapTypeBreakdown {
  gapType: string;
  count: number;
}

export interface SeverityBreakdown {
  severity: string;
  count: number;
}

export interface AnalysesSummary {
  totalAnalyses: number;
  totalGaps: number;
  avgGapsPerAnalysis: number;
  avgConfidence: number;
  gapTypeBreakdown: GapTypeBreakdown[];
  severityBreakdown: SeverityBreakdown[];
}

export interface FeedbackInput {
  isHelpful: boolean;
}

export interface Feedback {
  id: number;
  gapId: number;
  isHelpful: boolean;
  createdAt: string;
}