# lib/api-zod/src/generated/types/analysesSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRD Engineering Gap Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { GapTypeBreakdown } from './gapTypeBreakdown';
import type { SeverityBreakdown } from './severityBreakdown';

export interface AnalysesSummary {
  totalAnalyses: number;
  totalGaps: number;
  avgGapsPerAnalysis: number;
  avgConfidence: number;
  gapTypeBreakdown: GapTypeBreakdown[];
  severityBreakdown: SeverityBreakdown[];
}