# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRD Engineering Gap Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { Gap } from './gap';

export interface Analysis {
  id: number;
  prdText: string;
  title: string;
  createdAt: string;
  gaps: Gap[];
}