# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRD Engineering Gap Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisSummary {
  id: number;
  title: string;
  prdText: string;
  createdAt: string;
  gapCount: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
}