# lib/api-zod/src/generated/types/gap.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRD Engineering Gap Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface Gap {
  id: number;
  analysisId: number;
  gapType: string;
  description: string;
  severity: string;
  confidence: number;
  createdAt: string;
  helpfulCount: number;
  notHelpfulCount: number;
}