# lib/api-zod/src/generated/types/severityBreakdown.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PRD Engineering Gap Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface SeverityBreakdown {
  severity: string;
  count: number;
}